backend/requirements.txt
category-encoders==1.3.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
Deprecated==1.2.5
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.2
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.2
patsy==0.5.1
PyGithub==1.43.6
PyJWT==1.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
statsmodels==0.9.0
urllib3==1.24.1
Werkzeug==0.15.1
wrapt==1.11.1
